Change AEADCmd.{generateAEAD,validateAEAD} to using a byte array for data.
Retain String accepting versions of the API calls as well in the YubiHSM class.

It was discovered that you can't take any 8-bit data in a byte array and
turn it into a String that could then be turned into the same byte array
again when it was to be sent to the YubiHSM, since Java applies encoding
to strings and it was not apparent there was an 8-bit clean encoding.